refactor: Make headline service AntWeb-specific
